+++ public/assets/scripts/choices.search-prefix.js
@@ -1597,7 +1597,7 @@
             if (placeholderValue) {
                 inp.setAttribute('aria-label', placeholderValue);
             }
-            if (!labelId) {
+            else if (!labelId) {
                 addAriaLabel(this._docRoot, this.passedElement.element.id, inp);
             }
             return inp;
@@ -2457,12 +2457,18 @@
                 normalChoices.sort(config.sorter);
             }
             var choiceLimit = rendererableChoices.length;
+            var limit = choiceLimit;
             var sortedChoices = this._isSelectOneElement && placeholderChoices.length ? __spreadArray(__spreadArray([], placeholderChoices, true), normalChoices, true) : normalChoices;
             if (this._isSearching) {
-                choiceLimit = searchResultLimit;
+                if (searchResultLimit >= 0) {
+                    limit = searchResultLimit;
+                }
             }
-            else if (renderChoiceLimit && renderChoiceLimit > 0 && !withinGroup) {
-                choiceLimit = renderChoiceLimit;
+            else if (renderChoiceLimit >= 0 && !withinGroup) {
+                limit = renderChoiceLimit;
+            }
+            if (limit < choiceLimit) {
+                choiceLimit = limit;
             }
             // Add each choice to dropdown within range
             for (var i = 0; i < choiceLimit; i += 1) {